# Reduce allocations during fingerprinting.